test_roll_dice staticmethod made, passed all tests
