+++ ten_thousand/banker.py
@@ -1,0 +1,16 @@
+
+class Banker:
+    """docstring"""
+
+    def __init__(self):
+        pass
+
+    def shelf(self):  # input: amount of points(integer) to add to shelf
+        pass  # output: no return, temporarily stores unbanked points
+
+    def bank(self):  # takes shelf points and adds to total bank, resets shelf to 0
+        pass  # output: sum of bank and shelf
+
+    def clear_shelf(self):  # clear shelf points
+        pass
+
