+++ ten_thousand/game_logic.py
@@ -14,52 +14,26 @@ class GameLogic:
         pass
         # return integer score according to rules of game
 
-    @staticmethod  # input: integer between 1 and 6
-    def roll_dice(self, dice):
-        pass
-        # output: tuple with random values between 1 and 6
+    @staticmethod  # input: integer between 1 and 6, representing number of die
+    def roll_dice(dice):
+
+        if 1 <= dice <= 6:
+
+            dice_rack = []
+
+            for i in range(dice):
+                dice_value = random.randint(1, 6)
+                dice_rack.append(dice_value)
+            dice_tuple = tuple(dice_rack)
+            return dice_tuple
+
+        # output: tuple with random values between 1 and 6, representing fresh rolls
         # len() of tuple must match len() of input
 
 
-class Banker:
-    """docstring"""
+if __name__ == "__main__":
 
-    def __init__(self):
-        pass
+    obj = GameLogic
+    print(obj.roll_dice(4))
+    print(obj.roll_dice(6))
 
-    def shelf(self):  # input: amount of points(integer) to add to shelf
-        pass  # output: no return, temporarily stores unbanked points
-
-    def bank(self):  # takes shelf points and adds to total bank, resets shelf to 0
-        pass  # output: sum of bank and shelf
-
-    def clear_shelf(self):  # clear shelf points
-        pass
-
-
-
-
-
-
-
-
-
-dice_face = { 'one': 0,
-             'two': 0,
-             'three': 0,
-             'four': 0,
-             'five': 0,
-             'six': 0,
-             }
-# sides = dice_face.keys()
-sides = list(dice_face)
-
-for i in range(6):
-    dice_face[random.choice(sides)] += 1
-
-print('one:\t', dice_face['one'])
-print('two:\t', dice_face['two'])
-print('three:\t', dice_face['three'])
-print('four:\t', dice_face['four'])
-print('five:\t', dice_face['five'])
-print('six:\t', dice_face['six'])
